+++ event-to-sentry.go
@@ -16,6 +16,7 @@ import (
 	"cloud.google.com/go/storage"
 	"github.com/joho/godotenv"
 	_ "github.com/mattn/go-sqlite3"
+	"google.golang.org/api/iterator"
 )
 
 var httpClient = &http.Client{}
@@ -34,6 +35,11 @@ var (
 	projectDSNs map[string]*DSN
 	traceIds    []string
 )
+
+type Timestamp struct {
+	time.Time
+	rfc3339 bool
+}
 
 type DSN struct {
 	host      string
@@ -60,7 +66,7 @@ func parseDSN(rawurl string) *DSN {
 
 	var host string
 	if strings.Contains(rawurl, "ingest.sentry.io") {
-		// TODO need to slice the o87286 dynamically
+		// TODO slice the o87286 dynamically
 		host = "o87286.ingest.sentry.io"
 	}
 	if strings.Contains(rawurl, "@localhost:") {
@@ -113,44 +119,8 @@ func (d DSN) envelopeEndpoint() string {
 	return fullurl
 }
 
-// TODO put EventJson to its own eventjson.go
-type TypeSwitch struct {
-	Kind string `json:"type"`
-}
-type EventJson struct {
-	TypeSwitch `json:"type"`
-	*Error
-	*Transaction
-}
-
-func (eventJson *EventJson) UnmarshalJSON(data []byte) error {
-	if err := json.Unmarshal(data, &eventJson.TypeSwitch); err != nil {
-		return err
-	}
-	switch eventJson.Kind {
-	case "error":
-		eventJson.Error = &Error{}
-		return json.Unmarshal(data, eventJson.Error)
-	case "transaction":
-		eventJson.Transaction = &Transaction{}
-		return json.Unmarshal(data, eventJson.Transaction)
-	default:
-		return fmt.Errorf("unrecognized type value %q", eventJson.Kind)
-	}
-}
-
-type Event struct {
-	Platform string            `json:"platform"`
-	Kind     string            `json:"kind"`
-	Headers  map[string]string `json:"headers"`
-	Body     string            `json:"body"`
-}
-
-func (e Event) String() string {
-	return fmt.Sprintf("\n Event { Platform: %s, Type: %s }\n", e.Platform, e.Kind) // index somehow?
-}
+// TODO - could return an error instead
 func dsnToStoreEndpoint(projectDSNs map[string]*DSN, projectPlatform string) string {
-	// var storeEndpoint string
 	if projectPlatform == "javascript" {
 		return projectDSNs["javascript"].storeEndpoint()
 	} else if projectPlatform == "python" {
@@ -159,37 +129,6 @@ func dsnToStoreEndpoint(projectDSNs map[string]*DSN, projectPlatform string) str
 		log.Fatal("platform type not supported")
 	}
 	return ""
-}
-func matchDSN(projectDSNs map[string]*DSN, event Event) string {
-
-	platform := event.Platform
-
-	var storeEndpoint string
-	if platform == "javascript" && event.Kind == "error" {
-		storeEndpoint = projectDSNs["javascript"].storeEndpoint()
-	} else if platform == "python" && event.Kind == "error" {
-		storeEndpoint = projectDSNs["python"].storeEndpoint()
-	} else if platform == "android" && event.Kind == "error" {
-		storeEndpoint = projectDSNs["android"].storeEndpoint()
-	} else if platform == "javascript" && event.Kind == "transaction" {
-		storeEndpoint = projectDSNs["javascript"].envelopeEndpoint()
-	} else if platform == "python" && event.Kind == "transaction" {
-		storeEndpoint = projectDSNs["python"].envelopeEndpoint()
-	} else if platform == "android" && event.Kind == "transaction" {
-		storeEndpoint = projectDSNs["android"].envelopeEndpoint()
-	} else {
-		log.Fatal("platform type not supported")
-	}
-
-	if storeEndpoint == "" {
-		log.Fatal("missing store endpoint")
-	}
-	return storeEndpoint
-}
-
-type Timestamp struct {
-	time.Time
-	rfc3339 bool
 }
 
 func init() {
@@ -205,7 +144,6 @@ func init() {
 	py = flag.String("py", "", "python DSN")
 	flag.Parse()
 
-	// sentry +10.0.0 supports performance monitoring, transactions
 	projectDSNs = make(map[string]*DSN)
 	projectDSNs["javascript"] = parseDSN(os.Getenv("DSN_JAVASCRIPT_SAAS"))
 	if *js != "" {
@@ -224,87 +162,96 @@ func init() {
 }
 
 func main() {
-	// TODO read from CloudStorage
-	// jsonFile, err := os.Open(database)
-	// if err != nil {
-	// 	log.Fatal(err)
-	// }
 
-	readJsons(*ignore)
-	return
-
-	// CLOUD STORAGE
-	bucket := os.Getenv("BUCKET")
-	object := database
-	fmt.Println("DATASET object", object)
+	bucketName := os.Getenv("BUCKET")
+	// Initialize/Connect the Client
 	ctx := context.Background()
 	client, err := storage.NewClient(ctx)
 	if err != nil {
 		log.Fatalln("storage.NewClient:", err)
-		return
 	}
 	defer client.Close()
 	ctx, cancel := context.WithTimeout(ctx, time.Second*50)
 	defer cancel()
-	rc, err := client.Bucket(bucket).Object(object).NewReader(ctx)
-	if err != nil {
-		log.Fatalln("NewReader:", err)
-		return
-	}
-
-	// START
-	byteValue, _ := ioutil.ReadAll(rc) // jsonFile
-	// defer jsonFile.Close()
-	events := make([]Event, 0)
-	if err := json.Unmarshal(byteValue, &events); err != nil {
-		panic(err)
-	}
-	requests := []Transport{}
-	for _, event := range events {
-		fmt.Printf("\n> KIND|PLATFORM %v %v ", event.Kind, event.Platform)
-
-		if event.Kind == "error" {
-			bodyError, timestamper, bodyEncoder, storeEndpoint := decodeError(event)
-			bodyError = eventId(bodyError)
-			bodyError = release(bodyError)
-			bodyError = user(bodyError)
-			bodyError = timestamper(bodyError, event.Platform)
-
-			requests = append(requests, Transport{
-				kind:          event.Kind,
-				platform:      event.Platform,
-				eventHeaders:  event.Headers,
-				storeEndpoint: storeEndpoint,
-
-				bodyError:   bodyError,
-				bodyEncoder: bodyEncoder,
-			})
-
-		} else if event.Kind == "transaction" {
-			envelopeItems, envelopeTimestamper, envelopeEncoder, storeEndpoint := decodeEnvelope(event)
-			envelopeItems = eventIds(envelopeItems)
-			envelopeItems = envelopeTimestamper(envelopeItems, event.Platform)
-			envelopeItems = envelopeReleases(envelopeItems, event.Platform, event.Kind)
-			envelopeItems = removeLengthField(envelopeItems)
-			envelopeItems = sentAt(envelopeItems)
-			envelopeItems = users(envelopeItems)
-			getEnvelopeTraceIds(envelopeItems)
-
-			requests = append(requests, Transport{
-				kind:          event.Kind,
-				platform:      event.Platform,
-				eventHeaders:  event.Headers,
-				storeEndpoint: storeEndpoint,
-
-				envelopeItems:   envelopeItems,
-				envelopeEncoder: envelopeEncoder,
-			})
+	// Prepare bucket handle
+	bucketHandle := client.Bucket(bucketName)
+	// lists the contents of a bucket in Google Cloud Storage.
+	var fileNames []string
+	query := &storage.Query{Prefix: "eventtest"}
+	it := bucketHandle.Objects(ctx, query)
+	for {
+		obj, err := it.Next()
+		if err == iterator.Done {
+			break
 		}
-
+		if err != nil {
+			log.Fatalln("listBucket: unable to list bucket", err)
+		}
+		fileNames = append(fileNames, obj.Name)
+		printObj(obj)
 	}
 
-	setEnvelopeTraceIds(requests)
-	encodeAndSendEvents(requests, *ignore)
+	// TODO events.go could manage reading from storage. or like:
+	/*
+		storageClient := StorageClient(os.Getenv("BUCKET")) <-- is the init
+		//or
+		storageClient.init(os.Getenv("BUCKET"))
+		storageClient.query("event") .prefixQuery("event") .queryBucket .bucketQuery() .bucketSet()
+		storageClient.listBucketContents() .getBucket()
+		events := storageClient.getFiles() .bucketFiles()
+	*/
+
+	// Read each file's content
+	var events []EventJson
+	for _, fileName := range fileNames {
+		rc, err := bucketHandle.Object(fileName).NewReader(ctx)
+		if err != nil {
+			log.Fatalln("NewReader:", err)
+		}
+		byteValue, _ := ioutil.ReadAll(rc)
+		var event EventJson
+		// The EventJson's UnmarshalJSON overriden in event-to-sentry.go (soon EventJson.go)
+		if err := json.Unmarshal(byteValue, &event); err != nil {
+			panic(err)
+		}
+		events = append(events, event)
+	}
+
+	for _, event := range events {
+		if event.Kind == "error" {
+			event.Error.eventId()
+			event.Error.release()
+			event.Error.user()
+			event.Error.timestamp()
+		}
+		if event.Kind == "transaction" {
+			event.Transaction.eventId()
+			event.Transaction.release()
+			event.Transaction.user()
+			event.Transaction.timestamps()
+		}
+	}
+
+	getTraceIds(events)
+	updateTraceIds(events)
+
+	// TODO double check it's object was updated reference `fmt.Println("\n> timestamp AFTER", event.Error.Timestamp)`
+	requests := Requests{events}
+	requests.send()
 
 	return
+}
+
+func printObj(obj *storage.ObjectAttrs) {
+	fmt.Printf("filename: /%v/%v \n", obj.Bucket, obj.Name)
+	// fmt.Printf("ContentType: %q, ", obj.ContentType)
+	// fmt.Printf("ACL: %#v, ", obj.ACL)
+	// fmt.Printf("Owner: %v, ", obj.Owner)
+	// fmt.Printf("ContentEncoding: %q, ", obj.ContentEncoding)
+	// fmt.Printf("Size: %v, ", obj.Size)
+	// fmt.Printf("MD5: %q, ", obj.MD5)
+	// fmt.Printf("CRC32C: %q, ", obj.CRC32C)
+	// fmt.Printf("Metadata: %#v, ", obj.Metadata)
+	// fmt.Printf("MediaLink: %q, ", obj.MediaLink)
+	// fmt.Printf("StorageClass: %q, ", obj.StorageClass)
 }
